Define some command index numbers
These identify the operation that the board is to perform. Write
signifies that the board configuration is to change. Read only inputs
data from the board to the host.

Automatic processes like UVLO will not be controlled by the host.

The exactly on-the-wire format of these things is unimportant right now,
and can be normalized in the future.

Change-Id: Id751a13d7a7e3a7890b87e1351311229aa7a870b