Pump ADC current samples to a buffer
Samples read by the ADC will then be stored in a round-robin manner
whenever they get generated, into the buffer in output.c. The idea being
that we cycle through the output-driver ICs reading their currents.
Requestors for those values then read the most recently sampled value.
The max latency will be 4ms; it's another question as to whether this
is acceptable.

As of this commit, the current-sense-disable pin waggling has not been
implemented.

Change-Id: Id0e3bf1c73c74cf42663fe33bed697aff399ac57