Convert current sense report into milliamps
This is an initial approximation: the actual value reported by the ADC
is not linear (according to my observations) and not consistent between
outputs. Some kind of future settlement and data gathering needs to
occur, but in the meantime this is a good approximation to work with.

It's also an overapproximation, so safe for what we're doing

Change-Id: I541ae9ff64786300fcf68c6f8d70fcbe6b96c5a6