Write initial software current limit code
If we read a sample from the current measurement IC saying that we're
drawing more than 40A, cut all peripherals, and spin in a loop, beeping.

I can't test a current of 40A in my existing setup, but I'm sure this is
something we'll end up testing to qualify the firmware.

Change-Id: Ie9b8c24610cdcd85d53e7b1af56d057d1957db02