Disable current/voltage monitor for the moment
My current suspicion is that, with i2c clocked at something-Mhz (I
haven't figured out exactly what), the gap between each of these
functions completing the i2c transaction and starting the next one is
very small. Specifically, the hold time as observed by the INA219.

It demands 160nS bus downtime between reads, and we might not be giving
it to it. So, I'm going to dance around that for a while.

Change-Id: If3637d97847fe19e2a9f028faa811f239ca8506b