Disable battery polling again
After some experiments, if this block of code is enabled, then there's
a reasonably high chance of USB requests failing. With this code
enabled, the power board fails to enumerate 30% of the time. With it
disabled, it never fails.

I reason this is because of the amount of time spent blocked on i2c in
the polling loop (about 120uS every 250uS), as I don't think the
peripherals can be interacting in any way. To fix this, I'll be
switching i2c to be intr driven, not polling.

Change-Id: Iea234fb4d21ef83c421a58fac5a7f46f1f6b09de