Nix i2c interrupts
These actually aren't necessary: in all circumstances where the
peripheral needs some pieces of information that it doesn't have, the
clock gets stretched until it's available. (Just as the slave would do).

So we can in fact afford to run this whole thing in the polling loop.
And it works just fine. Hurrah, i2c hasn't destroyed my life (yet).

Change-Id: I2299125a8f5fc4ac919313a2852fda3fec79b5d4