Further adjust voltage iir
There's a very large voltage droop on startup if you have six motor
boards plugged in, due to inrush current. My power supply current limits
me, giving an inaccurate representation of what'll happen in the field.

On an actually battery, the droop recovers within 500uS, which is the
sample period of the INA219. The lowest it went was about 4V; if it goes
that low for one sample, with an average over 10 samples we'd be close
to tripping the UVLO logic. So, stick with a decay factor of 0.99 (or
100 samples, or whatever). Ultimately this is only to detect a constant
below-10.2v situation, which this will still achieve.

Change-Id: I7c6aa36c505a976d1754ac6a3b67c593ca42baf0