Configure watchdog to reset board after 50ms
Reset in the main loop on every iteration. If the board hangs for some
reason, we will fail safe. The 5v rail is not turned on until the
software UVLO has had ample opportunity to check the battery is not
flat, so there should be no scope for a continous power-on/reset cycle.

Change-Id: I9dedbebd2a8b0812ecdf0ed2149b905d51fd8a21