Reset watchdog while blocked in legitimate loops
There are two circumstances where we want to signal to the user while
blocking the main loop: over current, and under voltage. In those
circumstances we've already turned everything off, and we want to
continue signalling to the user that something is wrong without being
reset, so reset the watchdog.

Seeing how delay(2000) is used to time beeping and led flashing, which
is sufficient a pause (something like half a second) to trip the
watchdog, so split that delay up into a delay-and-watchdog-reset loop.

Change-Id: I29d93e01f623e80d7cd26c7d9330102b2dff37fc