Fix i2c bus hang caused by certain race conditions
Both i2c and USB are driven by polling on the power board, and as a
result heavy traffic on one can interfere with the timing / behaviour of
the other. During inital development I moved i2c to use a state machine
and not block on bus activity, as I was observing USB transaction
failures when the uC blocked on i2c transactions. Since then however
I've observed i2c failures during heavy USB traffic.

These failures presented themselves in two ways:
 * The i2c bus peripheral signalled 'BERR', bus error, and a voltage or
   current reading was aborted (due to software error handling).
 * 'ARLO', 'arbitration lost' was signalled. This wasn't handled by the
   software error handling, and all further i2c transactions blocked.

The root cause appears to be that, if the i2c fsm isn't executed in the
period between the 1st and 2nd data byte being read from the INA219,
then the 2nd byte is not NACK'd. Exactly what happens next is unclear,
but BERR and ARLO correspond to SDA transitioning when the peripheral
doesn't expect it.

The fix is to follow the application notes in the STM32F1x manual. That
says that the correct procedure is to set the 'POS' bit in the
peripheral, which makes the ACK bit apply to the /next/ byte, not the
current one. We then have a timing-safe read transaction:

 1 Start condition put on bus; clock stretches until fsm executes and
   supplies an address.
 2 Address put on bus; ADDR bit set if address acknowledged, clock
   stretches until fsm clears the ADDR bit
 3 At the same time we clear the ACK bit. The peripheral then reads
   one byte and acks it; reads a second bytes and nacks it; then
   stretches the clock until fsm executes and reads the data register
 4 We put a stop bit on the bus and wait until it's registered.

It seems that previously either 3 didn't stretch, or otherwise did
something crazy. It now no longer exhibits the two failure modes
described above.

Change-Id: I537ba6264e180f51bb9a74d21f838a72d3ce6c7a