Implement output current limit feature.
This installs one IIR per output, that decay over a period of 32
samples. Hitting 10A or 20A for low or high outputs repectively causes
them to be turned off, and the output status LED to be turned on.

Additionally, a flag is set indicating that the limit has occurred.
It can only be turned off by turning the status LED off; that can't be
controlled over USB. Only a board reset, overcurrent/undervolt, or USB
reset will cause the output to be re-enabled.

Change-Id: I590fe3f97b148518e2fb6a11e26929c6f736bdc7