Force SMPS on in bootloader
As referred to elsewhere, a design oversight is that if we perform a bad
flash of the power board, then it'll stick in the bootloader with the
odroid / brainboard turned off. At a normal school, nothing can get it
out of that state

Fix this by turning the 5v rail on in the bootloader. Advice from
Rob/Rich is that while this might not be ideal, we can rely on the
hardware UVLO to protect us from draining the battery in this
circumstance.

Change-Id: I308e318423a80cdd359e1860124ed505aaadca66