Decrease on-start wait time to one second
There's an initial delay to allow for battery/voltage monitors to
settle, during which we don't allow anything to happen. For some reason
I bumped it to 3 seconds last year, don't know why. Reducing it to one
second should be sufficient (IIRs take 100ms to build up).

There's the risk that we'll get battery-low reports when the battery is
actually sort-of sufficient. However I'm unable to create conditions
where this happens because the UVLO on my dev board kicks in at 10.8V.

Change-Id: I6d078c7d2fec10eff167d74b64ab21726a2e9e05